backend/tickets/serializers.py
from rest_framework import serializers

from .models import Ticket


class TicketSerializer(serializers.ModelSerializer):
    class Meta:
        model = Ticket
        fields = [
            "id",
            "title",
            "description",
            "category",
            "priority",
            "status",
            "created_at",
        ]
        read_only_fields = ["id", "created_at"]


class TicketPatchSerializer(serializers.ModelSerializer):
    """All fields optional for partial updates."""

    class Meta:
        model = Ticket
        fields = [
            "id",
            "title",
            "description",
            "category",
            "priority",
            "status",
            "created_at",
        ]
        read_only_fields = ["id", "created_at"]
        extra_kwargs = {
            "title": {"required": False},
            "description": {"required": False},
            "category": {"required": False},
            "priority": {"required": False},
            "status": {"required": False},
        }


class ClassifySerializer(serializers.Serializer):
    description = serializers.CharField(min_length=1)
