backend/requirements.txt
Django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
openai==1.59.5
dj-database-url==2.3.0
